output-parser-test-python/pyproject.toml
[project]
name = "output-parser-test-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.13",
    "langchain-core>=1.2.20",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
    "defusedxml>=0.7.1",
    "pymysql>=1.1.0",
    "cryptography>=46.0.5",
]
